Possible fix?
See this PR from scaffold-eth-2 for more details: https://github.com/scaffold-eth/scaffold-eth-2/pull/825
