+++ packages/nextjs/app/blockexplorer/address/[address]/page.tsx
@@ -1,9 +1,10 @@
-import fs from 'fs';
-import path from 'path';
-import { foundry } from 'viem/chains';
-import { AddressComponent } from '~~/app/blockexplorer/_components/AddressComponent';
-import deployedContracts from '~~/contracts/deployedContracts';
-import { GenericContractsDeclaration } from '~~/utils/scaffold-eth/contract';
+import fs from "fs";
+import path from "path";
+import { hardhat } from "viem/chains";
+import { AddressComponent } from "~~/app/blockexplorer/_components/AddressComponent";
+import deployedContracts from "~~/contracts/deployedContracts";
+import { isZeroAddress } from "~~/utils/scaffold-eth/common";
+import { GenericContractsDeclaration } from "~~/utils/scaffold-eth/contract";
 
 type PageProps = {
   params: { address: string };
@@ -11,13 +12,13 @@ type PageProps = {
 
 async function fetchByteCodeAndAssembly(buildInfoDirectory: string, contractPath: string) {
   const buildInfoFiles = fs.readdirSync(buildInfoDirectory);
-  let bytecode = '';
-  let assembly = '';
+  let bytecode = "";
+  let assembly = "";
 
   for (let i = 0; i < buildInfoFiles.length; i++) {
     const filePath = path.join(buildInfoDirectory, buildInfoFiles[i]);
 
-    const buildInfo = JSON.parse(fs.readFileSync(filePath, 'utf8'));
+    const buildInfo = JSON.parse(fs.readFileSync(filePath, "utf8"));
 
     if (buildInfo.output.contracts[contractPath]) {
       for (const contract in buildInfo.output.contracts[contractPath]) {
@@ -37,21 +38,21 @@ async function fetchByteCodeAndAssembly(buildInfoDirectory: string, contractPath
 
 const getContractData = async (address: string) => {
   const contracts = deployedContracts as GenericContractsDeclaration | null;
-  const chainId = foundry.id;
-  let contractPath = '';
+  const chainId = hardhat.id;
+  let contractPath = "";
 
   const buildInfoDirectory = path.join(
     __dirname,
-    '..',
-    '..',
-    '..',
-    '..',
-    '..',
-    '..',
-    '..',
-    'foundry',
-    'out',
-    'build-info',
+    "..",
+    "..",
+    "..",
+    "..",
+    "..",
+    "..",
+    "..",
+    "hardhat",
+    "artifacts",
+    "build-info",
   );
 
   if (!fs.existsSync(buildInfoDirectory)) {
@@ -76,8 +77,16 @@ const getContractData = async (address: string) => {
   return { bytecode, assembly };
 };
 
+export function generateStaticParams() {
+  // An workaround to enable static exports in Next.js, generating single dummy page.
+  return [{ address: "0x0000000000000000000000000000000000000000" }];
+}
+
 const AddressPage = async ({ params }: PageProps) => {
   const address = params?.address as string;
+
+  if (isZeroAddress(address)) return null;
+
   const contractData: { bytecode: string; assembly: string } | null = await getContractData(address);
   return <AddressComponent address={address} contractData={contractData} />;
 };
